Add lit-html client side template rendering
