+++ client_src/package.json
@@ -71,5 +71,8 @@
   "engine-strict": true,
   "engines": {
     "npm": ">= 6.10.0"
+  },
+  "dependencies": {
+    "lit-html": "^2.0.0"
   }
 }
